test: share config env helper
- move config env guard into utils::test_utils\n- update auth and command tests to use shared helper
